+++ backend/src/middlewares/authCheck.js
